requirements.txt
# FastAPI
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
asyncpg>=0.27.0 # Add asyncpg driver for PostgreSQL
# psycopg2-binary>=2.9.5 # Keep or remove depending if sync driver is still needed elsewhere
alembic>=1.9.0 # For migrations (good practice)

# Data Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Security
cryptography>=40.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography] # For potential JWT use later

# TabPFN Client
tabpfn-client>=0.1.7 # Or the appropriate version 